requirements.txt
alabaster==0.7.12
atomicwrites==1.4.0
attrs==21.2.0
Babel==2.9.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.2
colorama==0.4.4
distlib==0.3.2
docutils==0.17.1
filelock==3.0.12
idna==3.2
imagesize==1.2.0
imutils==0.5.4
iniconfig==1.1.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
numpy==1.21.0
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
packaging==21.0
platformdirs==2.0.2
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.18.0
pytz==2021.1
requests==2.26.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tk==0.1.0
toml==0.10.2
urllib3==1.26.6
virtualenv==20.6.0
